chatbot/requirements.txt
pymongo==4.3.3
pandas==1.5.3
python-dotenv==1.0.0
pydantic==1.10.8
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
langchain==0.0.184
openai==0.27.8
pydantic==1.10.8